Fixed the graphing analyzer so that it'd work with more than 1 sampler

+++ src/main/java/de/mvbonline/jmeteranalyzer/backend/graph/GraphAnalyzer.java
@@ -11,9 +11,7 @@ import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.Statement;
 import java.text.SimpleDateFormat;
-import java.util.Date;
-import java.util.List;
-import java.util.TimeZone;
+import java.util.*;
 
 /**
  * Created by mholz on 14.12.2015.
@@ -32,18 +30,22 @@ public class GraphAnalyzer implements Analyzer {
             long results = 0;
 
             while(true) {
-                StringBuilder jsData = new StringBuilder();
+                Map<String, StringBuilder> jsData = new HashMap<>();
                 String delim = "";
 
                 Date start = null;
                 Date end = null;
 
                 try(Statement statement = sqliteConnection.createStatement()) {
-                    String sql = "SELECT timestamp, duration, success FROM '" + table + "' ORDER BY timestamp ASC LIMIT " + PAGE_SIZE + " OFFSET " + results;
+                    String sql = "SELECT name, timestamp, duration, success FROM '" + table + "' ORDER BY timestamp ASC LIMIT " + PAGE_SIZE + " OFFSET " + results;
                     try(ResultSet resultSet = statement.executeQuery(sql)) {
                         boolean foundAny = false;
                         while(resultSet.next()) {
                             foundAny = true;
+                            String name = resultSet.getString("name");
+                            if(jsData.get(name) == null) {
+                                jsData.put(name, new StringBuilder());
+                            }
 
                             long timestamp = resultSet.getLong("timestamp");
                             if(start == null) {
@@ -51,23 +53,23 @@ public class GraphAnalyzer implements Analyzer {
                             }
                             end = new Date(timestamp);
 
-                            jsData.append(delim);
+                            jsData.get(name).append(delim);
                             delim = ",\n";
 
-                            jsData.append("{date: ");
-                            jsData.append(timestamp).append(", \n");
+                            jsData.get(name).append("{date: ");
+                            jsData.get(name).append(timestamp).append(", \n");
 
                             boolean success = resultSet.getBoolean("success");
                             if(success) {
-                                jsData.append("lineColor: '#5fb503',\n");
+                                jsData.get(name).append("lineColor: '#5fb503',\n");
                             } else {
-                                jsData.append("lineColor: '#ff0000',\n");
+                                jsData.get(name).append("lineColor: '#ff0000',\n");
                             }
 
-                            jsData.append("duration: ");
+                            jsData.get(name).append("duration: ");
 
                             long duration = resultSet.getLong("duration");
-                            jsData.append(duration).append("}");
+                            jsData.get(name).append(duration).append("}");
 
                             results++;
                         }
@@ -79,18 +81,25 @@ public class GraphAnalyzer implements Analyzer {
                 }
 
                 System.out.println("Writing page...");
-                try(InputStream is = GraphAnalyzer.class.getResourceAsStream("/graphs/graph.html")) {
-                    String file = IOUtils.toString(is);
-                    file = file.replace("$RESULT_DATA$", jsData.toString());
+                for (String s : jsData.keySet()) {
+                    try(InputStream is = GraphAnalyzer.class.getResourceAsStream("/graphs/graph.html")) {
+                        String file = IOUtils.toString(is);
+                        file = file.replace("$RESULT_DATA$", jsData.get(s).toString());
 
-                    SimpleDateFormat sdf = new SimpleDateFormat("yyyyMMdd-HHmmssSSS");
-                    sdf.setTimeZone(TimeZone.getTimeZone("GMT+00:00"));
-                    String startString = sdf.format(start);
-                    String endString = sdf.format(end);
+                        SimpleDateFormat sdf = new SimpleDateFormat("yyyyMMdd-HHmmssSSS");
+                        sdf.setTimeZone(TimeZone.getTimeZone("GMT+00:00"));
+                        String startString = sdf.format(start);
+                        String endString = sdf.format(end);
 
-                    File resultFile = new File(resultDir, table + "-" + startString + "-" + endString + ".duration-graph.html");
-                    try(OutputStream os = new FileOutputStream(resultFile)) {
-                        IOUtils.write(file, os);
+                        File subResultDir = new File(resultDir, "graphs/" + table + "/" + s);
+                        if(!subResultDir.exists() && !subResultDir.mkdirs()) {
+                            throw new IOException("Could not create dir " + subResultDir);
+                        }
+
+                        File resultFile = new File(subResultDir, startString + "-" + endString + ".duration-graph.html");
+                        try(OutputStream os = new FileOutputStream(resultFile)) {
+                            IOUtils.write(file, os);
+                        }
                     }
                 }
